requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
et_xmlfile==2.0.0
fastapi==0.110.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-auth==2.38.0
google-generativeai==0.3.2
googleapis-common-protos==1.69.1
grpcio==1.71.0rc2
grpcio-status==1.63.0rc1
h11==0.14.0
idna==3.10
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==10.2.0
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.6.1
pydantic_core==2.16.2
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
requests==2.31.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.27.1
wheel==0.45.1
